выпилил из мэйна запуск функции с параметрами
