+++ js/main.js
@@ -2,11 +2,8 @@ import {createDataPhotos} from'./picture/picture-data.js';
 import {creatingPictureElement} from'./picture/picture-rendering.js';
 import {pictureParameters} from'./picture/picture-parameters.js';
 
-import {createPictureFormEventHandlers} from'./picture_form/picture-form-events.js';
-import {pictureformParameters} from'./picture_form/picture-form-parameters.js';
-
+import './picture_form/picture-form.js';
 import './picture/picture-scale.js';
 import './picture/picture-effect.js';
 
 creatingPictureElement(createDataPhotos(),pictureParameters);
-createPictureFormEventHandlers(pictureformParameters);
